Initial window resizing.
Upscales the window context. Eventually we should check for aspect
ratio, but for now this is fine.

Started the wrapping of exposed third party headers from internal
headers. Eventually this should not expose the functionality but rather
should be wrapped in the correct "concept" with a suitable interface:

OpenGL -> Graphics
Imgui -> UI
etc.