Looking over an optimization for faces.
It is way incomplete at this moment.
Added some profiling about how much it takes to mesh it and I'm at ~0.25
ms, though it is on a extremelly powerful machine. But in any case, it
would seem the timing of the meshing is fine, considering it will occur
only on change, not every frame.