Improved timers and OpenGL handles.
Timers: They can be used to time start-end blocks, either to a log
system or manually. The first is used to trace the meshing timing and
the latter to time the application frame timing (~0.5 ms!).

The OpenGL handles are a owner handle wrapper that will free the
resource on destruction. It is templatized on the resource type so that
it calls the appropiate OpenGL releaser. It also does the right thing on
move and is not copiable. Hopefully this is all we need to not worry
anymore about OpenGL resource handling.